requirements.txt
aniso8601==8.1.1
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
protobuf==3.14.0
pytz==2021.1
six==1.15.0
uWSGI==2.0.19.1
Werkzeug==1.0.1
